SCR insolvency / Config and NFT take pool address
Previously, lendToPool was only considering the ocean, leaving always
the SCR. Now, was a parameter to consider all the total supply. This can
lead to a situation where SCR>totalSupply, in those cases, we call the
InsolvencyHook to give the opportunity to solve this lack of SCR.

Other change, now PolicyPoolConfig and PolicyNFT accept the address of
the PolicyPool in the initialization. This allows pre-calculate the
address of PolicyPool (with CREATE2) and create PolicyPoolConfig and
PolicyNFT with that address already connected.